requirements.txt
Flask
gunicorn
selenium
requests
python-dotenv
pandas
beautifulsoup4
lxml
markdown2
jinja2